Update Client.php
> Linkedin does give response but it's in the header not body. We should parse the response correctly.

Hi, I'm facing the same error. I think the problem is that the current responseToArray method are prepared to parse the body, but this api call, has no body in the response, it just has a new header with the ucgPost Id, as @soisme told.

I've done a little change in the method, so we can return the headers, but I'm not sure if this change could break any other api call, because I just started to use this package.